+++ command/secrets_tune_test.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/go-test/deep"
 	"github.com/hashicorp/vault/api"
+	"github.com/hashicorp/vault/helper/testhelpers/corehelpers"
 	"github.com/hashicorp/vault/sdk/helper/consts"
-	"github.com/hashicorp/vault/vault"
 	"github.com/mitchellh/cli"
 )
 
@@ -150,7 +150,7 @@ func TestSecretsTuneCommand_Run(t *testing.T) {
 	t.Run("integration", func(t *testing.T) {
 		t.Run("flags_all", func(t *testing.T) {
 			t.Parallel()
-			pluginDir, cleanup := vault.MakeTestPluginDir(t)
+			pluginDir, cleanup := corehelpers.MakeTestPluginDir(t)
 			defer cleanup(t)
 
 			client, _, closer := testVaultServerPluginDir(t, pluginDir)
