+++ vault/external_tests/raft/raft_test.go
@@ -23,6 +23,7 @@ import (
 	"github.com/hashicorp/vault/helper/constants"
 	"github.com/hashicorp/vault/helper/namespace"
 	"github.com/hashicorp/vault/helper/testhelpers"
+	"github.com/hashicorp/vault/helper/testhelpers/corehelpers"
 	"github.com/hashicorp/vault/helper/testhelpers/teststorage"
 	vaulthttp "github.com/hashicorp/vault/http"
 	"github.com/hashicorp/vault/internalshared/configutil"
@@ -236,7 +237,7 @@ func TestRaft_Retry_Join(t *testing.T) {
 			}
 
 			// Handle potential racy behavior with unseals. Retry the unseal until it succeeds.
-			vault.RetryUntil(t, 10*time.Second, func() error {
+			corehelpers.RetryUntil(t, 10*time.Second, func() error {
 				return cluster.AttemptUnsealCore(core)
 			})
 		}(t, clusterCore)
@@ -248,7 +249,7 @@ func TestRaft_Retry_Join(t *testing.T) {
 
 	vault.TestWaitActive(t, leaderCore.Core)
 
-	vault.RetryUntil(t, 10*time.Second, func() error {
+	corehelpers.RetryUntil(t, 10*time.Second, func() error {
 		return testhelpers.VerifyRaftPeers(t, cluster.Cores[0].Client, map[string]bool{
 			"core-0": true,
 			"core-1": true,
